show toast method를 utils 폴더로 이동
