+++ lib/core/constants/constants.dart
@@ -1,6 +1,6 @@
 class Constants {
 
   // aws public IP
-  static const baseUrl = "http://3.35.9.75";
+  static const baseUrl = "http://43.201.65.144";
 
 }
