+++ lib/presentation/search_screen/controller/search_controller.dart
@@ -1,7 +1,5 @@
 import '../../../data/controllers/heart_controller.dart';
-import '../../../data/repository/post_repository.dart';
 import '../../main_screen/controller/main_controller.dart';
-import '../../signup_screen/signup_screen.dart';
 import '/core/app_export.dart';
 import 'package:application4/presentation/search_screen/models/search_model.dart';
 import '../../../data/models/webtoon.dart';
